back-end/dashboard/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
gunicorn==20.1.0
django-cors-headers==4.3.1